Create base class for IsInside... functions
